refactor(apigateway): split data_plane.rs by concerns (3641 -> 2379 mod + 6 submodules)
Audit 2026-05-19 P9 remainder. data_plane.rs free-standing fns split:

- authorizers.rs   482  (enforce_authorizer / run_lambda_authorizer /
                        run_cognito_authorizer / cached-auth ladder /
                        gateway response overrides)
- integrations.rs  352  (http_proxy / vpc_link_proxy / mock_response /
                        aws_direct_integration + request/response template
                        application)
- usage_plans.rs   210  (enforce_usage_plan / parse_quota_period /
                        first_matching_plan / snapshot_plan / parse_throttle /
                        current_quota_window)
- validator.rs     128  (enforce_request_validator)
- routing.rs       115  (match_resource_path / extract_lambda_arn /
                        header_name_from_identity_source / stage_resource_arn /
                        extract_header_value / raw_token)
- errors.rs         51  (not_found / bad_gateway / unauthorized / forbidden /
                        api_key_forbidden / limit_exceeded / bad_request)

mod.rs retains the `handle` entry-point, ResolvedRoute helpers, the
TokenBucket impl, and the test corpus.

82 unit tests pass.
