+++ crates/fakecloud-apigateway/src/data_plane/mod.rs
@@ -447,945 +447,6 @@ pub async fn handle(
     result
 }
 
-fn not_found(msg: impl Into<String>) -> AwsServiceError {
-    AwsServiceError::aws_error(StatusCode::NOT_FOUND, "NotFoundException", msg.into())
-}
-
-fn bad_gateway(msg: impl Into<String>) -> AwsServiceError {
-    AwsServiceError::aws_error(StatusCode::BAD_GATEWAY, "BadGatewayException", msg.into())
-}
-
-fn unauthorized(msg: impl Into<String>) -> AwsServiceError {
-    AwsServiceError::aws_error(
-        StatusCode::UNAUTHORIZED,
-        "UnauthorizedException",
-        msg.into(),
-    )
-}
-
-fn forbidden(msg: impl Into<String>) -> AwsServiceError {
-    AwsServiceError::aws_error(StatusCode::FORBIDDEN, "AccessDeniedException", msg.into())
-}
-
-/// `ForbiddenException` matches the wire shape AWS returns for an
-/// API-key check failure (missing key / unknown key / disabled key).
-fn api_key_forbidden() -> AwsServiceError {
-    AwsServiceError::aws_error(StatusCode::FORBIDDEN, "ForbiddenException", "Forbidden")
-}
-
-/// `LimitExceededException` is the wire shape AWS uses when throttle or
-/// quota tripped at the data plane.
-fn limit_exceeded() -> AwsServiceError {
-    AwsServiceError::aws_error(
-        StatusCode::TOO_MANY_REQUESTS,
-        "LimitExceededException",
-        "Limit Exceeded",
-    )
-}
-
-/// Resolve the header name an authorizer's `identitySource` points at.
-/// AWS uses `method.request.header.<Name>` (e.g.
-/// `method.request.header.Authorization`). Bare names also work for
-/// callers that store just `Authorization`. Defaults to `Authorization`
-/// when nothing was configured.
-fn header_name_from_identity_source(source: Option<&str>) -> String {
-    let raw = source.unwrap_or("Authorization").trim();
-    // `identitySource` may contain comma-separated entries; use the
-    // first one (matches AWS's behaviour for primary-key caching).
-    let first = raw.split(',').next().unwrap_or(raw).trim();
-    if let Some(stripped) = first.strip_prefix("method.request.header.") {
-        stripped.to_string()
-    } else if first.is_empty() {
-        "Authorization".to_string()
-    } else {
-        first.to_string()
-    }
-}
-
-fn extract_header_value(req: &AwsRequest, header_name: &str) -> Option<String> {
-    req.headers
-        .get(header_name)
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.to_string())
-}
-
-/// Strip an optional `Bearer ` prefix from a TOKEN-authorizer header
-/// value before forwarding the raw token to the Lambda. AWS leaves the
-/// prefix in place, but Lambdas commonly receive it stripped too;
-/// preserve verbatim to match real behaviour.
-fn raw_token(value: &str) -> &str {
-    value
-}
-
-async fn enforce_authorizer(
-    service: &ApiGatewayService,
-    req: &AwsRequest,
-    api_id: &str,
-    stage: &str,
-    resource_path: &str,
-    authorization_type: &str,
-    authorizer: Option<&Authorizer>,
-) -> Result<Option<AuthorizerOutcome>, AwsServiceError> {
-    match authorization_type {
-        // Methods that haven't opted into authorization pass through.
-        // `AWS_IAM` is treated as "no method-level authorizer" here —
-        // SigV4 enforcement is handled upstream by the request signer.
-        "" | "NONE" | "AWS_IAM" => Ok(None),
-        "CUSTOM" | "TOKEN" | "REQUEST" => {
-            let authorizer = authorizer.ok_or_else(|| {
-                forbidden("Method requires a custom authorizer but none is configured")
-            })?;
-            run_lambda_authorizer(service, req, api_id, stage, resource_path, authorizer).await
-        }
-        "COGNITO_USER_POOLS" => {
-            let authorizer = authorizer.ok_or_else(|| {
-                forbidden("Method requires Cognito authorization but no authorizer is attached")
-            })?;
-            run_cognito_authorizer(service, req, authorizer).await
-        }
-        other => Err(forbidden(format!(
-            "Unsupported authorizationType '{other}'"
-        ))),
-    }
-}
-
-/// Invoke a TOKEN/REQUEST authorizer Lambda and translate its policy
-/// into an Allow/Deny outcome. Caches successful evaluations by
-/// `<authorizerId>|<token>` for `authorizerResultTtlInSeconds`.
-async fn run_lambda_authorizer(
-    service: &ApiGatewayService,
-    req: &AwsRequest,
-    api_id: &str,
-    stage: &str,
-    resource_path: &str,
-    authorizer: &Authorizer,
-) -> Result<Option<AuthorizerOutcome>, AwsServiceError> {
-    // For TOKEN authorizers AWS treats the value of the configured
-    // identity-source header as the cache key. For REQUEST authorizers
-    // the cache key concatenates all configured sources; we keep the
-    // simpler one-header model here because that's what real-world
-    // configurations use 95% of the time.
-    let header_name = header_name_from_identity_source(authorizer.identity_source.as_deref());
-    let token_value = extract_header_value(req, &header_name).ok_or_else(|| {
-        unauthorized(format!(
-            "Missing required identity source header '{header_name}'"
-        ))
-    })?;
-    if token_value.trim().is_empty() {
-        return Err(unauthorized(format!(
-            "Empty identity source header '{header_name}'"
-        )));
-    }
-
-    let cache_key = format!("{}|{}", authorizer.id, token_value);
-    if let Some(cached) = lookup_cached_auth(service, &req.account_id, &cache_key) {
-        return interpret_cached(cached);
-    }
-
-    let auth_uri = authorizer
-        .authorizer_uri
-        .as_deref()
-        .ok_or_else(|| bad_gateway("Authorizer is missing authorizerUri; cannot invoke Lambda"))?;
-    let function_arn = extract_lambda_arn(auth_uri)
-        .ok_or_else(|| bad_gateway("authorizerUri must reference a Lambda function ARN"))?;
-    let method_arn = build_method_arn(req, api_id, stage, resource_path);
-
-    let event = match authorizer.authorizer_type.as_str() {
-        "TOKEN" => {
-            json!({
-                "type": "TOKEN",
-                "methodArn": method_arn,
-                "authorizationToken": raw_token(&token_value),
-            })
-        }
-        // Default to REQUEST shape for any non-TOKEN Lambda authorizer.
-        _ => {
-            let mut headers = serde_json::Map::new();
-            for (k, v) in req.headers.iter() {
-                if let Ok(s) = v.to_str() {
-                    headers.insert(
-                        k.as_str().to_string(),
-                        serde_json::Value::String(s.to_string()),
-                    );
-                }
-            }
-            let mut query = serde_json::Map::new();
-            for (k, v) in &req.query_params {
-                query.insert(k.clone(), serde_json::Value::String(v.clone()));
-            }
-            json!({
-                "type": "REQUEST",
-                "methodArn": method_arn,
-                "resource": resource_path,
-                "path": req.raw_path,
-                "httpMethod": req.method.as_str(),
-                "headers": headers,
-                "queryStringParameters": query,
-                "requestContext": {
-                    "apiId": api_id,
-                    "stage": stage,
-                    "path": req.raw_path,
-                    "httpMethod": req.method.as_str(),
-                },
-            })
-        }
-    };
-
-    let delivery = service
-        .delivery()
-        .ok_or_else(|| bad_gateway("Lambda delivery not configured"))?;
-    let response_bytes = delivery
-        .invoke_lambda(&function_arn, &event.to_string())
-        .await
-        .ok_or_else(|| bad_gateway("Lambda delivery not configured"))?
-        .map_err(|e| forbidden(format!("Authorizer Lambda failed: {e}")))?;
-    let response: serde_json::Value = serde_json::from_slice(&response_bytes)
-        .map_err(|e| forbidden(format!("Authorizer returned invalid JSON: {e}")))?;
-
-    let principal_id = response
-        .get("principalId")
-        .and_then(|v| v.as_str())
-        .unwrap_or("user")
-        .to_string();
-    let context = response
-        .get("context")
-        .cloned()
-        .unwrap_or_else(|| serde_json::Value::Object(serde_json::Map::new()));
-    let effect = parse_policy_effect(&response, &method_arn);
-
-    let ttl = authorizer
-        .authorizer_result_ttl_in_seconds
-        .map(|v| v as i64)
-        .unwrap_or(DEFAULT_AUTHORIZER_TTL_SECS);
-    cache_auth_result(
-        service,
-        &req.account_id,
-        cache_key,
-        CachedAuthorizerResult {
-            principal_id: principal_id.clone(),
-            effect,
-            context: context.clone(),
-            claims: None,
-            expires_at: chrono::Utc::now() + chrono::Duration::seconds(ttl),
-        },
-    );
-
-    match effect {
-        AuthEffect::Allow => Ok(Some(AuthorizerOutcome {
-            principal_id,
-            context,
-            claims: None,
-        })),
-        AuthEffect::Deny => Err(forbidden("User is not authorized to access this resource")),
-    }
-}
-
-/// Walk `policyDocument.Statement` and resolve to a single Allow/Deny
-/// effect. Multiple matching Allow statements collapse to Allow; any
-/// Deny short-circuits to Deny (mirroring the IAM policy combinator).
-fn parse_policy_effect(response: &serde_json::Value, method_arn: &str) -> AuthEffect {
-    let Some(stmts) = response
-        .get("policyDocument")
-        .and_then(|p| p.get("Statement"))
-        .and_then(|s| s.as_array())
-    else {
-        return AuthEffect::Deny;
-    };
-    let mut allow = false;
-    for stmt in stmts {
-        let effect = stmt.get("Effect").and_then(|v| v.as_str()).unwrap_or("");
-        // Resource matching: explicit ARN match, wildcard `*`, or
-        // missing Resource (treat as `*`).
-        let matches = match stmt.get("Resource") {
-            Some(serde_json::Value::String(s)) => arn_matches(s, method_arn),
-            Some(serde_json::Value::Array(arr)) => arr
-                .iter()
-                .filter_map(|v| v.as_str())
-                .any(|s| arn_matches(s, method_arn)),
-            _ => true,
-        };
-        if !matches {
-            continue;
-        }
-        match effect {
-            "Deny" => return AuthEffect::Deny,
-            "Allow" => allow = true,
-            _ => {}
-        }
-    }
-    if allow {
-        AuthEffect::Allow
-    } else {
-        AuthEffect::Deny
-    }
-}
-
-/// Glob-match a policy resource expression (`arn:...:*` etc) against a
-/// concrete method ARN. `*` matches any sequence inside a single
-/// segment; `?` matches a single character.
-fn arn_matches(pattern: &str, target: &str) -> bool {
-    if pattern == "*" {
-        return true;
-    }
-    let mut p_chars = pattern.chars().peekable();
-    let mut t_chars = target.chars().peekable();
-    loop {
-        match (p_chars.peek().copied(), t_chars.peek().copied()) {
-            (None, None) => return true,
-            (None, Some(_)) => return false,
-            (Some('*'), _) => {
-                p_chars.next();
-                if p_chars.peek().is_none() {
-                    return true;
-                }
-                while t_chars.peek().is_some() {
-                    if arn_matches(
-                        &p_chars.clone().collect::<String>(),
-                        &t_chars.clone().collect::<String>(),
-                    ) {
-                        return true;
-                    }
-                    t_chars.next();
-                }
-                return false;
-            }
-            (Some('?'), Some(_)) => {
-                p_chars.next();
-                t_chars.next();
-            }
-            (Some(a), Some(b)) if a == b => {
-                p_chars.next();
-                t_chars.next();
-            }
-            _ => return false,
-        }
-    }
-}
-
-fn build_method_arn(req: &AwsRequest, api_id: &str, stage: &str, resource_path: &str) -> String {
-    let trimmed = resource_path.trim_start_matches('/');
-    format!(
-        "arn:aws:execute-api:{}:{}:{}/{}/{}/{}",
-        req.region,
-        req.account_id,
-        api_id,
-        stage,
-        req.method.as_str().to_uppercase(),
-        trimmed,
-    )
-}
-
-async fn run_cognito_authorizer(
-    service: &ApiGatewayService,
-    req: &AwsRequest,
-    authorizer: &Authorizer,
-) -> Result<Option<AuthorizerOutcome>, AwsServiceError> {
-    let header_name = header_name_from_identity_source(authorizer.identity_source.as_deref());
-    let token_value = extract_header_value(req, &header_name)
-        .ok_or_else(|| unauthorized(format!("Missing required JWT in header '{header_name}'")))?;
-    let token = token_value
-        .strip_prefix("Bearer ")
-        .or_else(|| token_value.strip_prefix("bearer "))
-        .unwrap_or(&token_value)
-        .trim();
-    if token.is_empty() {
-        return Err(unauthorized("Empty Authorization header"));
-    }
-
-    let cache_key = format!("{}|{}", authorizer.id, token);
-    if let Some(cached) = lookup_cached_auth(service, &req.account_id, &cache_key) {
-        return interpret_cached(cached);
-    }
-
-    let pool_arn = authorizer
-        .provider_arns
-        .first()
-        .ok_or_else(|| forbidden("Cognito authorizer has no providerARNs configured"))?;
-    let delivery = service
-        .delivery()
-        .ok_or_else(|| unauthorized("Cognito JWT verifier not configured"))?;
-    let claims = delivery
-        .verify_cognito_jwt(&req.account_id, pool_arn, token)
-        .map_err(|e| unauthorized(format!("Invalid JWT: {e}")))?;
-
-    let principal_id = claims
-        .get("sub")
-        .and_then(|v| v.as_str())
-        .unwrap_or("user")
-        .to_string();
-    let ttl = authorizer
-        .authorizer_result_ttl_in_seconds
-        .map(|v| v as i64)
-        .unwrap_or(DEFAULT_AUTHORIZER_TTL_SECS);
-    cache_auth_result(
-        service,
-        &req.account_id,
-        cache_key,
-        CachedAuthorizerResult {
-            principal_id: principal_id.clone(),
-            effect: AuthEffect::Allow,
-            context: serde_json::Value::Object(serde_json::Map::new()),
-            claims: Some(claims.clone()),
-            expires_at: chrono::Utc::now() + chrono::Duration::seconds(ttl),
-        },
-    );
-
-    Ok(Some(AuthorizerOutcome {
-        principal_id,
-        context: serde_json::Value::Object(serde_json::Map::new()),
-        claims: Some(claims),
-    }))
-}
-
-fn lookup_cached_auth(
-    service: &ApiGatewayService,
-    account_id: &str,
-    key: &str,
-) -> Option<CachedAuthorizerResult> {
-    let now = chrono::Utc::now();
-    let mut accounts = service.state_handle().write();
-    let state = accounts.get_or_create(account_id);
-    if let Some(cached) = state.authorizer_cache.get(key) {
-        if cached.expires_at > now {
-            return Some(cached.clone());
-        }
-    }
-    state.authorizer_cache.remove(key);
-    None
-}
-
-fn cache_auth_result(
-    service: &ApiGatewayService,
-    account_id: &str,
-    key: String,
-    entry: CachedAuthorizerResult,
-) {
-    let mut accounts = service.state_handle().write();
-    let state = accounts.get_or_create(account_id);
-    state.authorizer_cache.insert(key, entry);
-}
-
-fn interpret_cached(
-    cached: CachedAuthorizerResult,
-) -> Result<Option<AuthorizerOutcome>, AwsServiceError> {
-    match cached.effect {
-        AuthEffect::Allow => Ok(Some(AuthorizerOutcome {
-            principal_id: cached.principal_id,
-            context: cached.context,
-            claims: cached.claims,
-        })),
-        AuthEffect::Deny => Err(forbidden("User is not authorized to access this resource")),
-    }
-}
-
-/// Apply the customer-configured gateway response template (if any) for
-/// `response_type`. Returns `None` when no override is registered, in
-/// which case the caller should propagate the original `AwsServiceError`
-/// unchanged. AWS allows overriding the HTTP status code and the
-/// response body via `responseTemplates` keyed by content type; we honor
-/// both and substitute `$context.error.messageString` /
-/// `$context.error.responseType` so the standard AWS-recommended template
-/// `{"message":$context.error.messageString}` renders correctly.
-fn apply_gateway_response_override(
-    service: &ApiGatewayService,
-    account_id: &str,
-    api_id: &str,
-    response_type: &str,
-    err: &AwsServiceError,
-) -> Option<AwsResponse> {
-    let accounts = service.state_handle().read();
-    let state = accounts.get(account_id)?;
-    let value = state.gateway_responses.get(api_id)?.get(response_type)?;
-    // `statusCode` may be a string or numeric per AWS docs; accept both
-    // and reject anything that doesn't fit a u16 instead of silently
-    // truncating it.
-    let status_code = value
-        .get("statusCode")
-        .and_then(|v| {
-            v.as_str().and_then(|s| s.parse::<u16>().ok()).or_else(|| {
-                v.as_u64()
-                    .filter(|n| *n <= u16::MAX as u64)
-                    .map(|n| n as u16)
-            })
-        })
-        .and_then(|n| StatusCode::from_u16(n).ok())
-        .unwrap_or_else(|| err.status());
-    let templates = value.get("responseTemplates").and_then(|v| v.as_object());
-    let template = templates
-        .and_then(|t| t.get("application/json").and_then(|v| v.as_str()))
-        .map(|s| s.to_string());
-    let body = match template {
-        Some(t) => render_error_template(&t, response_type, &err.message()),
-        // Default body matches AWS's built-in shape for an UNAUTHORIZED /
-        // ACCESS_DENIED response.
-        None => format!("{{\"message\":\"{}\"}}", escape_json(&err.message())),
-    };
-    Some(AwsResponse {
-        status: status_code,
-        content_type: "application/json".to_string(),
-        body: bytes::Bytes::from(body.into_bytes()).into(),
-        headers: http::HeaderMap::new(),
-    })
-}
-
-/// Substitute the two `$context.error.*` variables AWS exposes in
-/// gateway response templates. Anything else is left verbatim — full VTL
-/// rendering belongs to integration request/response transforms, not
-/// here.
-fn render_error_template(template: &str, response_type: &str, message: &str) -> String {
-    let escaped = escape_json(message);
-    template
-        .replace("$context.error.messageString", &format!("\"{escaped}\""))
-        .replace("$context.error.message", &escaped)
-        .replace("$context.error.responseType", response_type)
-}
-
-fn escape_json(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    for ch in s.chars() {
-        match ch {
-            '"' => out.push_str("\\\""),
-            '\\' => out.push_str("\\\\"),
-            '\n' => out.push_str("\\n"),
-            '\r' => out.push_str("\\r"),
-            '\t' => out.push_str("\\t"),
-            c if (c as u32) < 0x20 => out.push_str(&format!("\\u{:04x}", c as u32)),
-            c => out.push(c),
-        }
-    }
-    out
-}
-
-fn inject_authorizer_into_event(event: &mut serde_json::Value, outcome: &AuthorizerOutcome) {
-    let Some(req_ctx) = event
-        .get_mut("requestContext")
-        .and_then(|v| v.as_object_mut())
-    else {
-        return;
-    };
-    let mut auth_obj = serde_json::Map::new();
-    auth_obj.insert(
-        "principalId".to_string(),
-        serde_json::Value::String(outcome.principal_id.clone()),
-    );
-    if let serde_json::Value::Object(ctx) = &outcome.context {
-        for (k, v) in ctx {
-            auth_obj.insert(k.clone(), v.clone());
-        }
-    }
-    if let Some(claims) = &outcome.claims {
-        auth_obj.insert("claims".to_string(), claims.clone());
-    }
-    req_ctx.insert(
-        "authorizer".to_string(),
-        serde_json::Value::Object(auth_obj),
-    );
-}
-
-/// Match `template` (e.g. `/items/{id}/parts`) against `path_segments`
-/// (e.g. `["items", "42", "parts"]`). Returns the path-parameter map on
-/// match, or `None` if the path doesn't fit the template.
-fn match_resource_path(
-    template: &str,
-    path_segments: &[String],
-) -> Option<BTreeMap<String, String>> {
-    // Root resource: only an empty (or missing) remaining path matches.
-    if template == "/" {
-        return if path_segments.is_empty() {
-            Some(BTreeMap::new())
-        } else {
-            None
-        };
-    }
-    let template_segments: Vec<&str> = template.split('/').filter(|s| !s.is_empty()).collect();
-    let mut params = BTreeMap::new();
-    let mut t = 0;
-    let mut p = 0;
-    while t < template_segments.len() {
-        let seg = template_segments[t];
-        if seg.starts_with('{') && seg.ends_with('}') {
-            let inner = seg.trim_start_matches('{').trim_end_matches('}');
-            if let Some(name) = inner.strip_suffix('+') {
-                // Greedy match — consume the remainder.
-                if p >= path_segments.len() {
-                    return None;
-                }
-                params.insert(name.to_string(), path_segments[p..].join("/"));
-                return Some(params);
-            }
-            if p >= path_segments.len() {
-                return None;
-            }
-            params.insert(inner.to_string(), path_segments[p].clone());
-            p += 1;
-        } else {
-            if p >= path_segments.len() || path_segments[p] != seg {
-                return None;
-            }
-            p += 1;
-        }
-        t += 1;
-    }
-    if p == path_segments.len() {
-        Some(params)
-    } else {
-        None
-    }
-}
-
-/// Pull the function ARN out of an AWS_PROXY integration URI of the
-/// shape `arn:aws:apigateway:<region>:lambda:path/<api-version>/functions/<arn>/invocations`.
-fn extract_lambda_arn(uri: &str) -> Option<String> {
-    if !uri.contains(":lambda:path/") {
-        return None;
-    }
-    let prefix = uri.split("/functions/").nth(1)?;
-    let arn = prefix.trim_end_matches("/invocations");
-    Some(arn.to_string())
-}
-
-async fn http_proxy(
-    req: &AwsRequest,
-    integration: &Integration,
-    body_override: Option<bytes::Bytes>,
-) -> Result<AwsResponse, AwsServiceError> {
-    let url = integration
-        .uri
-        .as_ref()
-        .ok_or_else(|| bad_gateway("HTTP integration missing uri"))?;
-    let method = match req.method {
-        Method::GET => reqwest::Method::GET,
-        Method::POST => reqwest::Method::POST,
-        Method::PUT => reqwest::Method::PUT,
-        Method::DELETE => reqwest::Method::DELETE,
-        Method::PATCH => reqwest::Method::PATCH,
-        Method::HEAD => reqwest::Method::HEAD,
-        Method::OPTIONS => reqwest::Method::OPTIONS,
-        _ => reqwest::Method::GET,
-    };
-    let client = reqwest::Client::new();
-    let mut builder = client.request(method, url);
-    for (k, v) in req.headers.iter() {
-        if let Ok(s) = v.to_str() {
-            builder = builder.header(k.as_str(), s);
-        }
-    }
-    let body = body_override.as_ref().unwrap_or(&req.body);
-    if !body.is_empty() {
-        builder = builder.body(body.clone().to_vec());
-    }
-    let resp = builder
-        .send()
-        .await
-        .map_err(|e| bad_gateway(format!("backend HTTP failure: {e}")))?;
-    let status = StatusCode::from_u16(resp.status().as_u16()).unwrap_or(StatusCode::BAD_GATEWAY);
-    let mut headers = http::HeaderMap::new();
-    for (k, v) in resp.headers().iter() {
-        if let (Ok(name), Ok(val)) = (
-            http::HeaderName::from_bytes(k.as_str().as_bytes()),
-            http::HeaderValue::from_bytes(v.as_bytes()),
-        ) {
-            // `append` preserves multi-value headers like multiple
-            // `Set-Cookie` lines that the backend may emit.
-            headers.append(name, val);
-        }
-    }
-    let content_type = headers
-        .get("content-type")
-        .and_then(|v| v.to_str().ok())
-        .unwrap_or("application/octet-stream")
-        .to_string();
-    let body = resp
-        .bytes()
-        .await
-        .map_err(|e| bad_gateway(format!("backend body read failure: {e}")))?;
-    Ok(AwsResponse {
-        status,
-        content_type,
-        headers,
-        body: bytes::Bytes::from(body.to_vec()).into(),
-    })
-}
-
-/// Resolve a VPC_LINK integration by looking up the VpcLink's
-/// `targetArns`, finding the first NLB/ALB that has a bound port in
-/// the ELBv2 dataplane, and forwarding the request there.
-async fn vpc_link_proxy(
-    req: &AwsRequest,
-    integration: &Integration,
-    service: &ApiGatewayService,
-) -> Result<AwsResponse, AwsServiceError> {
-    let connection_id = integration
-        .connection_id
-        .as_deref()
-        .ok_or_else(|| bad_gateway("VPC_LINK integration missing connectionId"))?;
-
-    let target_arns: Vec<String> = {
-        let accounts = service.state_handle().read();
-        let state = accounts
-            .get(&req.account_id)
-            .ok_or_else(|| bad_gateway("VPC_LINK: account not found in API Gateway state"))?;
-        let vpc_link = state
-            .vpc_links
-            .get(connection_id)
-            .ok_or_else(|| bad_gateway(format!("VPC_LINK not found: {connection_id}")))?;
-        let target_arns = vpc_link
-            .get("targetArns")
-            .and_then(|v| v.as_array())
-            .ok_or_else(|| bad_gateway("VPC_LINK missing targetArns"))?;
-        if target_arns.is_empty() {
-            return Err(bad_gateway("VPC_LINK targetArns is empty"));
-        }
-        target_arns
-            .iter()
-            .filter_map(|v| v.as_str().map(|s| s.to_string()))
-            .collect()
-    };
-
-    let elbv2 = service
-        .elbv2_state
-        .as_ref()
-        .ok_or_else(|| bad_gateway("VPC_LINK not available: ELBv2 state not wired"))?;
-    let port = {
-        let elbv2_read = elbv2.read();
-        let elbv2_account = elbv2_read
-            .get(&req.account_id)
-            .ok_or_else(|| bad_gateway("VPC_LINK: account not found in ELBv2 state"))?;
-        let mut bound_port = None;
-        for arn in &target_arns {
-            if let Some(lb) = elbv2_account.load_balancers.get(arn) {
-                if let Some(port) = lb.bound_port {
-                    bound_port = Some(port);
-                    break;
-                }
-            }
-        }
-        bound_port.ok_or_else(|| {
-            bad_gateway("VPC_LINK: none of the target NLBs/ALBs have an active dataplane port")
-        })?
-    };
-
-    // Build the backend URL from the integration URI, replacing the
-    // host with the local ELBv2 dataplane endpoint.
-    let original_url = integration.uri.as_deref().unwrap_or("http://localhost/");
-    let path_and_query = http::Uri::try_from(original_url)
-        .ok()
-        .and_then(|u| u.path_and_query().map(|p| p.as_str().to_string()))
-        .unwrap_or_else(|| req.raw_path.clone());
-    let backend_url = format!("http://127.0.0.1:{port}{path_and_query}");
-
-    let mut proxy_integration = integration.clone();
-    proxy_integration.uri = Some(backend_url);
-    http_proxy(req, &proxy_integration, None).await
-}
-
-/// Apply the integration's `requestTemplates` to the request body.
-/// Returns `Some(transformed_body)` when a template matched the request
-/// content type, or `None` to leave the body unchanged.
-fn apply_request_template(
-    req: &AwsRequest,
-    integration: &Integration,
-    vtl_ctx: &mut crate::vtl::Context,
-) -> Option<bytes::Bytes> {
-    let content_type = req
-        .headers
-        .get("content-type")
-        .and_then(|v| v.to_str().ok())
-        .unwrap_or("application/json");
-    let normalized = content_type
-        .split(';')
-        .next()
-        .unwrap_or(content_type)
-        .trim();
-    let template = integration.request_templates.get(normalized)?;
-    let rendered = crate::vtl::render(template, vtl_ctx);
-    Some(bytes::Bytes::from(rendered))
-}
-
-/// Apply the integration response's `responseTemplates` to the backend
-/// response body. Returns a new `AwsResponse` with the rendered template.
-#[allow(clippy::too_many_arguments)]
-async fn apply_response_template(
-    backend_resp: AwsResponse,
-    integration: &Integration,
-    req: &AwsRequest,
-    vtl_ctx: &mut crate::vtl::Context,
-    api_id: &str,
-    resource_path: &str,
-    _stage_name: &str,
-    _path_params: &BTreeMap<String, String>,
-    _stage_vars: &BTreeMap<String, String>,
-    service: &ApiGatewayService,
-) -> Result<AwsResponse, AwsServiceError> {
-    let status_code = backend_resp.status.as_u16().to_string();
-    let key = response_key(
-        api_id,
-        resource_path,
-        &integration.http_method,
-        &status_code,
-    );
-    let accounts = service.state_handle().read();
-    let state = accounts.get(&req.account_id);
-    let resp_template = state.and_then(|st| {
-        st.integration_responses
-            .get(&key)
-            .and_then(|v| v.get("responseTemplates"))
-            .and_then(|t| t.as_object())
-    });
-    let content_type = backend_resp
-        .content_type
-        .as_str()
-        .split(';')
-        .next()
-        .unwrap_or(&backend_resp.content_type)
-        .trim();
-    let template = resp_template
-        .and_then(|t| t.get(content_type))
-        .and_then(|v| v.as_str());
-    let body = if let Some(template) = template {
-        // Inject $input with the backend response body so the template
-        // can reference it.
-        let body_str = String::from_utf8_lossy(backend_resp.body.expect_bytes()).to_string();
-        let body_json: Value = serde_json::from_str(&body_str).unwrap_or(Value::Null);
-        vtl_ctx.set("input", json!({"body": body_str, "json": body_json}));
-        crate::vtl::render(template, vtl_ctx).into_bytes()
-    } else {
-        backend_resp.body.expect_bytes().to_vec()
-    };
-    Ok(AwsResponse {
-        status: backend_resp.status,
-        content_type: backend_resp.content_type,
-        headers: backend_resp.headers,
-        body: bytes::Bytes::from(body).into(),
-    })
-}
-
-/// Build a MOCK integration response by looking up the integration
-/// response for the method and rendering its `responseTemplates`.
-#[allow(clippy::too_many_arguments)]
-async fn mock_response(
-    req: &AwsRequest,
-    integration: &Integration,
-    vtl_ctx: &mut crate::vtl::Context,
-    api_id: &str,
-    resource_path: &str,
-    _stage_name: &str,
-    service: &ApiGatewayService,
-) -> Result<AwsResponse, AwsServiceError> {
-    let method = integration.http_method.as_str();
-    // Default to 200 when no explicit status code is configured.
-    let default_status = "200";
-    let key = response_key(api_id, resource_path, method, default_status);
-    let accounts = service.state_handle().read();
-    let state = accounts.get(&req.account_id);
-    // Try the 200 response first; if absent scan for any integration
-    // response registered for this method and use its configured status.
-    let resp_record = state.and_then(|st| {
-        st.integration_responses.get(&key).or_else(|| {
-            let prefix = format!("{api_id}/{resource_path}/{method}/");
-            st.integration_responses
-                .iter()
-                .find(|(k, _)| k.starts_with(&prefix))
-                .map(|(_, v)| v)
-        })
-    });
-    let (status, resp_templates) = if let Some(record) = resp_record {
-        let status = record
-            .get("statusCode")
-            .and_then(|v| v.as_str())
-            .unwrap_or(default_status);
-        let templates = record.get("responseTemplates").and_then(|v| v.as_object());
-        (status, templates)
-    } else {
-        (default_status, None)
-    };
-    let status = status
-        .parse::<u16>()
-        .ok()
-        .and_then(|n| StatusCode::from_u16(n).ok())
-        .unwrap_or(StatusCode::OK);
-    let content_type = "application/json";
-    let body = if let Some(templates) = resp_templates {
-        templates
-            .get(content_type)
-            .and_then(|v| v.as_str())
-            .map(|t| crate::vtl::render(t, vtl_ctx))
-            .unwrap_or_default()
-    } else {
-        String::new()
-    };
-    Ok(AwsResponse {
-        status,
-        content_type: content_type.to_string(),
-        headers: http::HeaderMap::new(),
-        body: bytes::Bytes::from(body).into(),
-    })
-}
-
-/// Dispatch an `AWS` direct service integration to the corresponding
-/// fakecloud service handler. The integration URI follows the API Gateway
-/// format: `arn:aws:apigateway:{region}:{service}:action/{Action}` or
-/// `arn:aws:apigateway:{region}:{service}:path/{path}`.
-async fn aws_direct_integration(
-    req: &AwsRequest,
-    uri: &str,
-    service: &ApiGatewayService,
-) -> Result<AwsResponse, AwsServiceError> {
-    let registry = service.registry().ok_or_else(|| {
-        bad_gateway("AWS direct integration not available: service registry not wired")
-    })?;
-    let registry = registry.get().ok_or_else(|| {
-        bad_gateway("AWS direct integration not available: service registry not yet populated")
-    })?;
-
-    let parts: Vec<&str> = uri.split(':').collect();
-    if parts.len() < 6 || parts[0] != "arn" || parts[1] != "aws" || parts[2] != "apigateway" {
-        return Err(bad_gateway(format!(
-            "AWS integration uri not in expected ARN format: {uri}"
-        )));
-    }
-    let target_service = parts[4];
-    let action_or_path = parts[5];
-
-    let target = registry.get(target_service).ok_or_else(|| {
-        bad_gateway(format!(
-            "AWS integration target service '{target_service}' not registered"
-        ))
-    })?;
-
-    let mut dispatch_req = AwsRequest {
-        service: target_service.to_string(),
-        action: req.action.clone(),
-        region: req.region.clone(),
-        account_id: req.account_id.clone(),
-        request_id: uuid::Uuid::new_v4().to_string(),
-        headers: req.headers.clone(),
-        query_params: req.query_params.clone(),
-        body: req.body.clone(),
-        body_stream: parking_lot::Mutex::new(None),
-        path_segments: req.path_segments.clone(),
-        raw_path: req.raw_path.clone(),
-        raw_query: req.raw_query.clone(),
-        method: req.method.clone(),
-        is_query_protocol: false,
-        access_key_id: req.access_key_id.clone(),
-        principal: req.principal.clone(),
-    };
-
-    if let Some(action) = action_or_path.strip_prefix("action/") {
-        dispatch_req.action = action.to_string();
-    } else if let Some(path) = action_or_path.strip_prefix("path/") {
-        dispatch_req.raw_path = format!("/{path}");
-        dispatch_req.path_segments = path.split('/').map(|s| s.to_string()).collect();
-    } else {
-        return Err(bad_gateway(format!(
-            "AWS integration uri must contain action/ or path/ segment: {uri}"
-        )));
-    }
-
-    target.handle(dispatch_req).await
-}
-
 // ── Usage plan throttle + quota ──
 
 /// In-memory throttle + quota state. Buckets are keyed by
@@ -1445,96 +506,6 @@ impl TokenBucket {
     }
 }
 
-/// Resolve the usage plan that matches `(api_id, stage_name)` for an
-/// API key. If no plan matches, the request is unmetered (the key is
-/// known but not associated with a plan that targets this stage).
-fn enforce_usage_plan(
-    service: &ApiGatewayService,
-    req: &AwsRequest,
-    api_id: &str,
-    stage_name: &str,
-    method_path: &str,
-) -> Result<(), AwsServiceError> {
-    let presented = req
-        .headers
-        .get("x-api-key")
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.to_string());
-    let presented = match presented {
-        Some(s) if !s.trim().is_empty() => s,
-        _ => return Err(api_key_forbidden()),
-    };
-
-    // Resolve key + the active usage plan under a single read lock.
-    // `plan = None` when the key isn't associated with any plan that
-    // targets this stage; in that case the request is unmetered.
-    let (key_id, plan): (String, Option<UsagePlanSnapshot>) = {
-        let accounts = service.state_handle().read();
-        let state = accounts
-            .get(&req.account_id)
-            .ok_or_else(api_key_forbidden)?;
-        let key = state
-            .api_keys
-            .values()
-            .find(|k| k.value == presented)
-            .cloned();
-        let key = match key {
-            Some(k) if k.enabled => k,
-            _ => return Err(api_key_forbidden()),
-        };
-        let plan = first_matching_plan(state, &key.id, api_id, stage_name, method_path);
-        (key.id, plan)
-    };
-    let Some(plan) = plan else {
-        return Ok(());
-    };
-
-    let mut meters = service.meters.lock();
-
-    // Throttle. A method-level override (`apiStages[].throttle[<path>]`)
-    // takes precedence over the plan-level throttle. The bucket key
-    // includes the method override path so each `(plan, key, method)`
-    // triple meters independently — matching AWS's documented behavior
-    // that method-level limits are evaluated separately from the plan
-    // overall throttle.
-    if let Some((rate, burst, method_override)) = plan.throttle {
-        let bucket_key = (
-            req.account_id.clone(),
-            plan.id.clone(),
-            key_id.clone(),
-            method_override.unwrap_or_default(),
-        );
-        let bucket = meters
-            .buckets
-            .entry(bucket_key)
-            .or_insert_with(|| TokenBucket::new(rate, burst));
-        // Keep config fresh in case the plan was updated.
-        bucket.rate_per_sec = rate;
-        bucket.burst = burst;
-        if !bucket.try_acquire(Instant::now()) {
-            return Err(limit_exceeded());
-        }
-    }
-
-    // Quota.
-    if let Some((limit, period, offset)) = plan.quota {
-        let window = current_quota_window(chrono::Utc::now(), period, offset);
-        let counter_key = (
-            req.account_id.clone(),
-            plan.id.clone(),
-            key_id.clone(),
-            window,
-        );
-        let entry = meters.counters.entry(counter_key).or_insert(0);
-        if *entry >= limit {
-            return Err(limit_exceeded());
-        }
-        *entry += 1;
-    }
-
-    Ok(())
-}
-
 #[derive(Debug, Clone)]
 struct UsagePlanSnapshot {
     id: String,
@@ -1555,252 +526,6 @@ enum QuotaPeriod {
     Day,
     Week,
     Month,
-}
-
-fn parse_quota_period(s: &str) -> Option<QuotaPeriod> {
-    match s {
-        "DAY" => Some(QuotaPeriod::Day),
-        "WEEK" => Some(QuotaPeriod::Week),
-        "MONTH" => Some(QuotaPeriod::Month),
-        _ => None,
-    }
-}
-
-/// AWS picks any matching plan when multiple plans associate the same
-/// key; we pick the first per `BTreeMap` iteration order, which is
-/// deterministic by `usagePlanId`. `method_path` (e.g. `/items/GET`)
-/// drives selection of the per-method throttle override under the
-/// matched `apiStages[]` entry.
-fn first_matching_plan(
-    state: &crate::state::ApiGatewayState,
-    key_id: &str,
-    api_id: &str,
-    stage_name: &str,
-    method_path: &str,
-) -> Option<UsagePlanSnapshot> {
-    for (plan_id, keys) in &state.usage_plan_keys {
-        if !keys.contains_key(key_id) {
-            continue;
-        }
-        let Some(plan) = state.usage_plans.get(plan_id) else {
-            continue;
-        };
-        let matched_stage = plan.api_stages.iter().find(|stage_entry| {
-            let api = stage_entry.get("apiId").and_then(|v| v.as_str());
-            let stage = stage_entry.get("stage").and_then(|v| v.as_str());
-            matches!((api, stage), (Some(a), Some(s)) if a == api_id && s == stage_name)
-        });
-        let Some(matched_stage) = matched_stage else {
-            continue;
-        };
-        return Some(snapshot_plan(plan, matched_stage, method_path));
-    }
-    None
-}
-
-fn snapshot_plan(
-    plan: &crate::state::UsagePlan,
-    matched_stage: &serde_json::Value,
-    method_path: &str,
-) -> UsagePlanSnapshot {
-    let plan_throttle = plan.throttle.as_ref().and_then(parse_throttle);
-    // Method-level override under the matched apiStage entry. AWS keys
-    // these by `<resource_path>/<HTTP_METHOD>`. We try the exact path
-    // first, then the AWS catch-all `/*/*` if no exact match.
-    let method_throttle = matched_stage
-        .get("throttle")
-        .and_then(|t| t.as_object())
-        .and_then(|map| {
-            let exact = map.get(method_path);
-            let wildcard = map.get("/*/*");
-            exact
-                .or(wildcard)
-                .and_then(parse_throttle)
-                .map(|(rate, burst)| (rate, burst, Some(method_path.to_string())))
-        });
-    let throttle =
-        method_throttle.or_else(|| plan_throttle.map(|(rate, burst)| (rate, burst, None)));
-    let quota = plan.quota.as_ref().and_then(|q| {
-        let limit = q.get("limit").and_then(|v| v.as_u64())?;
-        let period_str = q.get("period").and_then(|v| v.as_str())?;
-        let period = parse_quota_period(period_str)?;
-        let offset = q.get("offset").and_then(|v| v.as_i64()).unwrap_or(0);
-        Some((limit, period, offset))
-    });
-    UsagePlanSnapshot {
-        id: plan.id.clone(),
-        throttle,
-        quota,
-    }
-}
-
-/// Parse a `{rateLimit, burstLimit}` JSON object into a `(rate, burst)`
-/// pair. Returns `None` for missing fields, non-numeric values, or
-/// non-positive limits — AWS treats those as "no throttle configured".
-fn parse_throttle(t: &serde_json::Value) -> Option<(f64, f64)> {
-    let rate = t.get("rateLimit").and_then(|v| v.as_f64())?;
-    let burst = t
-        .get("burstLimit")
-        .and_then(|v| v.as_f64().or_else(|| v.as_i64().map(|n| n as f64)))?;
-    if rate <= 0.0 || burst <= 0.0 {
-        return None;
-    }
-    Some((rate, burst))
-}
-
-/// Compute the canonical window key for a quota period. AWS resets
-/// quota counters at the start of the period (UTC midnight for DAY,
-/// Sunday for WEEK, first-of-month for MONTH); `offset` shifts the
-/// boundary by N days. The returned string is purely a counter key.
-#[allow(clippy::too_many_arguments)]
-fn enforce_request_validator(
-    service: &ApiGatewayService,
-    req: &AwsRequest,
-    api_id: &str,
-    _stage_name: &str,
-    validator_id: &str,
-    request_parameters: &BTreeMap<String, bool>,
-    request_models: &BTreeMap<String, String>,
-    path_params: &BTreeMap<String, String>,
-) -> Result<(), AwsServiceError> {
-    let accounts = service.state_handle().read();
-    let state = accounts
-        .get(&req.account_id)
-        .ok_or_else(|| bad_request("Request validation failed"))?;
-
-    let validator = state
-        .request_validators
-        .get(api_id)
-        .and_then(|m| m.get(validator_id))
-        .ok_or_else(|| bad_request("Request validation failed"))?;
-
-    let validate_body = validator
-        .get("validateRequestBody")
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-    let validate_params = validator
-        .get("validateRequestParameters")
-        .and_then(|v| v.as_bool())
-        .unwrap_or(false);
-
-    if validate_params {
-        for (param_key, required) in request_parameters {
-            if !*required {
-                continue;
-            }
-            let present = if let Some(name) = param_key.strip_prefix("method.request.querystring.")
-            {
-                req.query_params
-                    .get(name)
-                    .map(|v| !v.is_empty())
-                    .unwrap_or(false)
-            } else if let Some(name) = param_key.strip_prefix("method.request.path.") {
-                path_params
-                    .get(name)
-                    .map(|v| !v.is_empty())
-                    .unwrap_or(false)
-            } else if let Some(name) = param_key.strip_prefix("method.request.header.") {
-                req.headers
-                    .get(name)
-                    .and_then(|v| v.to_str().ok())
-                    .map(|v| !v.trim().is_empty())
-                    .unwrap_or(false)
-            } else {
-                // Unknown parameter source — skip.
-                true
-            };
-            if !present {
-                return Err(bad_request(format!(
-                    "Missing required request parameter: {}",
-                    param_key
-                )));
-            }
-        }
-    }
-
-    if validate_body {
-        let content_type = req
-            .headers
-            .get("content-type")
-            .and_then(|v| v.to_str().ok())
-            .unwrap_or("application/json");
-        // Normalize content-type by stripping charset suffix.
-        let normalized = content_type
-            .split(';')
-            .next()
-            .unwrap_or(content_type)
-            .trim();
-
-        let model_name = match request_models.get(normalized) {
-            Some(name) => name,
-            None => match request_models.get("$default") {
-                Some(name) => name,
-                None => {
-                    // No model for this content type and no $default — skip body validation.
-                    return Ok(());
-                }
-            },
-        };
-
-        let model = state
-            .models
-            .get(api_id)
-            .and_then(|m| m.get(model_name))
-            .ok_or_else(|| bad_request("Request body does not match model schema"))?;
-
-        let schema_str = model
-            .schema
-            .as_deref()
-            .ok_or_else(|| bad_request("Request body does not match model schema"))?;
-        let schema: serde_json::Value = serde_json::from_str(schema_str)
-            .map_err(|_| bad_request("Request body does not match model schema"))?;
-
-        let body_json: serde_json::Value = serde_json::from_slice(&req.body)
-            .map_err(|_| bad_request("Request body does not match model schema"))?;
-
-        drop(accounts);
-
-        if let Err(e) = crate::model_validation::validate(&schema, &body_json) {
-            return Err(bad_request(format!(
-                "Request body does not match model schema for content type {}: {}",
-                normalized, e
-            )));
-        }
-    }
-
-    Ok(())
-}
-
-fn bad_request(msg: impl Into<String>) -> AwsServiceError {
-    AwsServiceError::aws_error(StatusCode::BAD_REQUEST, "BadRequestException", msg.into())
-}
-
-fn current_quota_window(
-    now: chrono::DateTime<chrono::Utc>,
-    period: QuotaPeriod,
-    offset: i64,
-) -> String {
-    use chrono::Datelike;
-    let date = now.date_naive() - chrono::Duration::days(offset);
-    match period {
-        QuotaPeriod::Day => format!("D:{date}"),
-        QuotaPeriod::Week => {
-            // ISO week: Monday-based. AWS docs aren't specific about
-            // anchor day; ISO week is stable + locale-free.
-            let iso = date.iso_week();
-            format!("W:{}-{:02}", iso.year(), iso.week())
-        }
-        QuotaPeriod::Month => format!("M:{}-{:02}", date.year(), date.month()),
-    }
-}
-
-// ─── WAFv2 inspection ──────────────────────────────────────────────
-
-/// Build the resource ARN that callers use when associating a WebACL
-/// with an API Gateway v1 stage:
-/// `arn:aws:apigateway:<region>::/restapis/<api>/stages/<stage>`.
-fn stage_resource_arn(region: &str, api_id: &str, stage_name: &str) -> String {
-    format!("arn:aws:apigateway:{region}::/restapis/{api_id}/stages/{stage_name}",)
 }
 
 /// Run the WAFv2 evaluator for one API Gateway v1 request. Returns
@@ -3639,3 +2364,16 @@ mod tests {
         assert_eq!(dispatched.path_segments, vec![""]); // path/ splits to [""]
     }
 }
+
+mod authorizers;
+mod errors;
+mod integrations;
+mod routing;
+mod usage_plans;
+mod validator;
+use authorizers::*;
+use errors::*;
+use integrations::*;
+use routing::*;
+use usage_plans::*;
+use validator::*;
